Made number of kernel evals a long
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/trunk@11613 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ weka/src/main/java/weka/classifiers/functions/supportVector/RegOptimizer.java
@@ -120,7 +120,7 @@ public class RegOptimizer implements OptionHandler, Serializable,
   protected SMOset m_supportVectors;
 
   /** number of kernel evaluations, used for printing statistics only **/
-  protected int m_nEvals = 0;
+  protected long m_nEvals = 0;
 
   /** number of kernel cache hits, used for printing statistics only **/
   protected int m_nCacheHits = -1;
@@ -266,7 +266,7 @@ public class RegOptimizer implements OptionHandler, Serializable,
    * 
    * @return the number of kernel evaluations
    */
-  public int getKernelEvaluations() {
+  public long getKernelEvaluations() {
     return m_nEvals;
   }
